+++ riverpie/CHANGELOG.md
@@ -3,6 +3,7 @@
 - feat: notifiers expose a `provider` getter to access the corresponding provider
 - fix: `ref.dispose` should also dispose dependencies
 - fix: change generics of method signature of `ReduxAction.external` to keep the state type
+- fix: handle case when action is finished after the notifier is disposed
 
 ## 0.27.0